moved ui to brave shields
