+++ browser/brave_shields/android/ephemeral_storage_utils.cc
@@ -6,8 +6,8 @@
 #include <cstddef>
 
 #include "base/android/jni_android.h"
+#include "brave/browser/brave_shields/android/jni_headers/BraveFirstPartyStorageCleanerUtils_jni.h"
 #include "brave/browser/brave_shields/brave_shields_tab_helper.h"
-#include "brave/browser/ephemeral_storage/android/jni_headers/BraveFirstPartyStorageCleanerUtils_jni.h"
 #include "brave/browser/ephemeral_storage/ephemeral_storage_tab_helper.h"
 #include "chrome/browser/android/tab_android.h"
