+++ browser/brave_shields/android/java/src/org/chromium/chrome/browser/brave_shields/FirstPartyStorageCleanerInterface.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.ephemeral_storage;
+package org.chromium.chrome.browser.brave_shields;
 
 import org.chromium.build.annotations.NullMarked;
